refactor the location of class
