+++ src/main/java/seedu/scheduler/logic/commands/AddCommand.java
@@ -13,7 +13,6 @@ import static seedu.scheduler.logic.parser.CliSyntax.PREFIX_VENUE;
 
 import java.util.Collections;
 
-import seedu.scheduler.commons.web.ConnectToGoogleCalendar;
 import seedu.scheduler.logic.CommandHistory;
 import seedu.scheduler.logic.RepeatEventGenerator;
 import seedu.scheduler.logic.commands.exceptions.CommandException;
