+++ src/main/java/seedu/scheduler/logic/commands/DeleteCommand.java
@@ -8,7 +8,6 @@ import java.util.List;
 import seedu.scheduler.commons.core.Messages;
 import seedu.scheduler.commons.core.index.Index;
 import seedu.scheduler.commons.util.EventFormatUtil;
-import seedu.scheduler.commons.web.ConnectToGoogleCalendar;
 import seedu.scheduler.logic.CommandHistory;
 import seedu.scheduler.logic.commands.exceptions.CommandException;
 import seedu.scheduler.logic.parser.Flag;
